Show Kimu link only in public signup form
This replaces the :new macro by helper calls for maintainability.
